login

+++ src/myboard/controller/DeleteServlet.java
@@ -8,6 +8,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 /**
@@ -21,11 +22,17 @@ public class DeleteServlet extends HttpServlet{
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         request.setCharacterEncoding("UTF-8");
 
-        int id = Integer.parseInt(request.getParameter("id"));
-        String pw = request.getParameter("pw");
-        boardRepository.deleteBoard(id, pw);
+        // 로그인 체크하여 비로그인시 로그인창으로 이동
+        HttpSession session = request.getSession();
+        if(session.getAttribute("isLogin") == null) {
+            response.sendRedirect("/board/loginForm");
+        } else {
+            int id = Integer.parseInt(request.getParameter("id"));
+            String pw = request.getParameter("pw");
+            boardRepository.deleteBoard(id, pw);
 
-        //list로 이동
-        response.sendRedirect("/board/list");
+            //list로 이동
+            response.sendRedirect("/board/list");
+        }
     }
 }
